apps/agent/server/requirements.txt
psutil>=5.9.0
requests>=2.28.0
python-dotenv>=1.0.0
pynvml>=11.0.0
